ca-certificates handler - run on change at end of role
Handlers are the correct solution to this type of problem but
its difficult to get just this handler to run at the end of the role.
Instead, provide a mechanism to execute the handler if the role changed.

This will result in double handler execution but I'm not sure that can
be avoided at this stage.

Note that the handler name has been simplified to "update ca-certificates"

Fixes #3